leitor_definitivo(códigos python)/requirements.txt
opencv-python
pillow
pytesseract
paddleocr
mysql-connector-python
tk
